Changed to using long double for storing free memory to avoid problems on 64bit systems. Also made changed to the Windows code, making it probably not work on Win9x

+++ cc.c
@@ -1163,7 +1163,7 @@ void memory_update(struct _instance *inst, HWND owner)
 	if(inst && inst->pixmap && *(inst->pixmap))
 	{
 		TextConfig *texts = (TextConfig *)inst->custom;
-		unsigned long Memory;
+		long double Memory;
 
 		Get_Memory(&Memory);
 		ScaledPrint(&texts->text1[9], Memory, 2);
